refactor syllabi_service
