+++ .gitignore
@@ -22,3 +22,7 @@ __pycache__/
 python/logs/
 .DS_Store/
 experiments/llm-evaluation/venv
+.venv/
+.pytest_cache/
+.idea/
+python/services/logs/
