+++ python/services/syllabi_service/app/core/__init__.py
@@ -1,0 +1,1 @@
+
