+++ python/services/syllabi_service/app/main.py
@@ -1,0 +1,6 @@
+import uvicorn
+
+if __name__ == "__main__":
+    config = uvicorn.Config("app.api.routes:app", host="127.0.0.1", port=8001, reload=True)
+    server = uvicorn.Server(config)
+    server.run()
