+++ python/services/syllabi_service/app/schema/__init__.py
@@ -1,0 +1,1 @@
+
