+++ python/services/syllabi_service/app/schema/courseSyllabi.py
