+++ python/services/syllabi_service/tests/data/syllabi/TEST101_2025W1_PDF_Tabular_Assessments.pdf
