+++ python/services/syllabi_service/tests/data/syllabi/TEST101_2025W1_PDF_Without_Bullet_CLOs.pdf
