+++ python/services/syllabi_service/tests/data/syllabi/TEST101_2025W1_Word_Assessments_with_brackets.docx
